Initial work towards reworking the DCC signal generation
This commit splits the DCC Signal code into three parts:
1) programming interface
2) common signal code
3) hardware timer specific implementation

This is a precursor to implementing a new signal generator to solve a
big bug in the LCC integration (crash when track power is ON).
