+++ include/DCCSignalGenerator_Timer.h
@@ -1,0 +1,33 @@
+/**********************************************************************
+DCC COMMAND STATION FOR ESP32
+
+COPYRIGHT (c) 2017-2019 Mike Dunston
+
+  This program is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+  You should have received a copy of the GNU General Public License
+  along with this program.  If not, see http://www.gnu.org/licenses
+**********************************************************************/
+#pragma once
+
+#include "DCCSignalGenerator.h"
+#include <esp32-hal-timer.h>
+
+class SignalGenerator_HardwareTimer : public SignalGenerator {
+public:
+  SignalGenerator_HardwareTimer(String, uint16_t, uint8_t, uint8_t);
+  // these need to be public for the ISR handler to access them
+  hw_timer_t *_timer;
+  bool _topOfWave{true};
+protected:
+  void enable() override;
+  void disable() override;
+  void lockSendQueueISR() override;
+  void unlockSendQueueISR() override;
+};
